+++ packages/fuselage-hooks/src/useStableArray.ts
@@ -1,14 +1,9 @@
 import { useRef, MutableRefObject } from 'react';
 
-/**
- * @hidden
- */
-type Comparator = <T>(a: T, b: T) => boolean;
-
 const hasChanged = <T extends unknown[]>(
   ref: MutableRefObject<T>,
   array: T,
-  compare: Comparator,
+  compare: <T>(a: T, b: T) => boolean,
 ): boolean => {
   if (!Array.isArray(array)) {
     return true;
@@ -28,7 +23,7 @@ const hasChanged = <T extends unknown[]>(
 const getCurrentArray = <T extends unknown[]>(
   ref: MutableRefObject<T>,
   array: T,
-  compare: Comparator,
+  compare: <T>(a: T, b: T) => boolean,
 ): T => {
   if (hasChanged(ref, array, compare)) {
     ref.current = array;
@@ -45,8 +40,9 @@ const getCurrentArray = <T extends unknown[]>(
  *        equal
  * @returns the passed array if the elements are NOT equals; the previously
  *          stored array otherwise
+ * @public
  */
-export const useStableArray = <T extends unknown[]>(array: T, compare: Comparator = Object.is): T => {
+export const useStableArray = <T extends unknown[]>(array: T, compare: <T>(a: T, b: T) => boolean = Object.is): T => {
   const ref = useRef(Array.isArray(array) ? array : [] as T);
   return getCurrentArray<T>(ref, array, compare);
 };